WIP: move default mapping store proeprty to dataset layer config - undo moving mapping selection to a shared component
