+++ frontend/javascripts/types/schemas/dataset_view_configuration.schema.ts
@@ -14,6 +14,7 @@ export function getDefaultLayerViewConfiguration(
     isDisabled: false,
     isInverted: false,
     isInEditMode: false,
+    defaultMapping: null,
   };
   return { ...defaultLayerViewConfiguration, ...dynamicDefault };
 }
